bugfix #1
Drawing a cached character was causing an error because the image that was being drawn had a height of 0. Tracking down the source of the NaNs lead me to camera updating, the delta between frames becomes large when you switch tabs and come back later. The unusually large delta causes an Infinity to divide by itself, which evaluates to NaN.